Truncate 'time' field of measurements to nearest month, close #87
Also switch ichnaea.tasks.trim_excessive_data to use 'created'
to judge the amount of time a measure has been in the DB, rather
than 'time'. This condition is used to keep records around long
enough to get summarized into the stats tables, so 'created' is
the correct field to consider; using 'time' was an unreported bug.