Fix stat adjustment in invalid lac migration, merge the two deletes.
The stat adjustment was missing the 65534 gsm family case and the
update statement used a hardcoded key=2. To be really correct we also
have to filter out cells which were added today, as those didn't yet
make it into the stat numbers. Finally made the stat select and delete
statement the same.