Use apolitical names on country stats page, refs Bugzilla 1172833.
Specifially this uses `Palestine`, `Taiwan` and `Holy See` without
any further words to describte their political status. We also
update some of the language on the statistic pages to refer to
countries or regions, to make it clear we don't make claims about
the political status of any of the listed entries.