Maintain a region code estimate for new wifi networks, refs #222.
This is a 80% solution and uses the bounding box data to match the
aggregated lat/lon against all bounding boxes. If the station fits
only into a single box, we pick that country/region.
We'll see what percentage of networks this approach
covers and how many times we get no or multiple answers.