Further improve using signal strength in location queries.
We use a non-linear least squares optimization instead of a simple
weighted average. Also use each station's score to select the best
sample of networks and use it in the initial guess. Finally estimate
the accuracy of the result as the 95th percentile of the distance to
all the stations. We don't use the station radius, as our radius
estimates are fairly bad bounding box estimates.