Add a new `location_dump` command line script to dump/export networks.
This exports data from any of the sharded station tables to a local
gzipped CSV file with data from all columns, including internal ones.

The script optionally supports specifying an area restriction to only
export data from a certain area. The area is specified as a center point
(latitude + longitude) and a radius. Internally a bounding box is used,
so slightly more than the desired circular area is being exported.