Use separate fallback cache namespaces for different provider schemas.
This means if one where to use both unwiredlabs and a second ichnaea
instance as fallbacks for different keys, their contents would no
longer be cached in the same namespace and influence each other.
On the other hand caches are still shared on a per schema basis, so
two API keys with different fallback names and different tokens
would still share the same cache contents.