get the containers building again
this patch adds a specific step for installing gnupg ahead of adding the mysql apt repo and then rerunning apt-get update and install for the remaining dependencies. It's a bit of a hack but it will do for now.

since pip has no real dependency ordering, we also need to shuffle around the order in which we install some of the libraries. this patch doesn't incorporate library changes but the next step will require changing the mysql connector lib, which will require this change.