update mysql-connector to mysql-connector-python
since the last build of master the mysql-connector libs were deleted, reuploaded with incorrect version numbers, and eventually the project was handed over to oracle. Oracle deprecated the whole thing and asked consumers to switch to mysql-connector-python, which has a downgrade (that fails) or an upgrade from 2.x to 8.x. See: https://github.com/pypa/warehouse/issues/5537

this patch upgrades to the latest available version of mysql-connector-python. The default behavior of the library changes in 8.0.11 in a way that causes existing tests to vomit  on UnicodeDecodeErrors coming out of the CEXT. Rather than dig into why this is happening (https://stackoverflow.com/a/55150960), this patch opts for the minimally invasive route of restoring the pure python default of the 2.x library we upgraded from. A conditional statement in the database class adds the appropriate argument when the mysqlconnector transport lib is in use.