update the database
testing the healthcheck system requires running the health check, and one of those checks verifies the freshness of the geoip database. this patch updates test data, which is pulled from the source-data and test-data folders of the repo described by the readme in the test data folder.

this also updates the maxmind library and version, though in the future when this needs to be updated that isn't strictly necessary