Document requirements, split constraints
Determine the primary requirements, using pipdeptree and looking for
import statements in the code. For each primary requirement, add a
one-line summary for each, and link to the source repo. Move the
requirements of the primary requirements to a new constraints.txt file.

Reformat and complete the hash list for each with hashin.