Fix "make test" so it works
I changed the pytest params to settings that are like what we did in
Socorro. It makes it easier to see when modules are skipped and to get a
feel for how long tests take to run and what's going on.

I moved the pytest configuration to setup.cfg so we have one less file
floating around.