Rework tests scaffolding to work in CircleCI
Circle CI can't load volumes through docker-compose, so we have to
create a data container and then mount that in the container running the
tests.

I also changed how the local dev environment test works so that it uses
the same machinery--that way we're running the tests in the same way
that Circle CI runs them.