Remove DB_TYPE default
Previously, the default for DB_TYPE was the ddl uri, so if you didn't
specify a type or forgot to specify a type when using configure_db(),
you'd get the ddl one with admin privs to the db. That seems like a bad
default. I did a pass where I changed the default to the readonly uri.

In this PR, I further change it so there's no default at all. I also
went through the codebase and fixed instances of create_db() and
drop_db() to always use the ddl uri and configure_db() to always pass a
type.