Redo map_tiles related things
MAP_TILES_URL and MAP_TILES_SRC were getting computed as module
constants, but they depend on ASSET_URL which got messy in the tests.

This redoes that so they're no longer module constants. It switches one
to be a function and the other gets folded into get_csp_policy().
Further, it sets ASSET_URL in test.env so it's always set that way and
we don't have to monkeypatch it.