Adjust strategy for INSERT collisions
When inserting rows, an error is raised if a primary or unique key is
duplicated. The code attempts to avoid this and use an UPDATE instead.
It also uses a custom SQL compiler rule to insert an "ON DUPLICATE KEY
UPDATE" clause, a MySQL-specific syntax that supports update-or-insert
operations.

When the intent was to update-or-insert, use .on_duplicate_key_update,
added by SQLAlchemy 1.2.

When the intent was to ignore the insert on a duplicate key, change
the INSERT to an INSERT IGNORE using .prefix_with("IGNORE").

Remove the custom compile rule "on_duplicate" and the related tests.