Fix database ping in webapp startup
Previously, a Database.ping() method was used to test the database
connection when configuring the webapp. This was removed in a previous
PR because it appeared to be unused, based on a code search and tests
passing.

Instead of bringing .ping() back for a single caller, add a function
ichnaea.db.ping_session(), which can be used to test the connectivity of
a database session. This is now used when configuring the webapp and
when calling the __heartbeat__ endpoint.