Update gunicorn, gevent, and greenlet
This updates the gunicorn, gevent, and greenlet dependencies.

This also removes gevent monkey patching from the tests--it wasn't
getting run early enough so it was emitting a warning and the comment
suggested it should only monkey patch the webapp tests. We can do that,
but if we do do that, we need to separate the webapp tests into a
separate test suite. Until we do that, I think it's best to remove the
monkey patching altogether.

Fixes #873