Migrate simplejson to json
Replace simplejson library with json standard library. The interface is
nearly identical. For example, loads() returns str not bytes for both
libraries in Python 3.

Drop the "internal_json" serializer based on simplejson and used by
Celery. Set Celery options to just "json", which is also the default in
Celery 4.

The requests library will prefer simplejson if available, so it must be
removed (make build) or tests will fail.