Remove mysqlconnector requirement and usage
mysqlconnector was added a couple of years ago for the datamap and dump
scripts. However, it adds a lot of complexity to the codebase and forces
us to keep tabs on two different mysql libraries.

A lot of things have changed in the last couple of years in the world of
Python and mysql clients and we think the compelling reasons to have two
are gone now. Thus we're going to simplify the situation and switch back
to PyMySQL for all mysql connections.

Fixes #870