update_statregion: Use ORM, not bulk_*_mappings
MySQL returns the number of rows affected by an UPDATE, and
bulk_update_mappings raises a StaleDataError if the affected rows are
less than the rows sent.  This means that the update fails if the radio
count is the same for any region. Instead, return to using the ORM
methods, which sends several UPDATEs, but only when the data changes.