Switch to Markus for metrics
This removes the statsd-related scaffolding in favor of the Markus
library. Markus makes it easier to configure a project for local
development differently than server environments. It also comes with
testing tools for asserting things about generated metrics.

This removes the statsd-related scaffolding and replaces it with
equivalent Markus things.