Force pytest assert rewriting in tests/base.py
Tell pytest to rewrite assertions in base.py, so that we have more
information about the occasional test failures. This is needed because
these files don't match the pattern test_*.py or *_test.py. However,
since there is just two of them in the code base, this targeted change
is a better solution than modifying the configuration python_files or
changing the file names.