Test CellAreaUpdater with region=None
Add CellAreaUpdater tests for cases where cell regions are None,
meaning that a region could not be determined from the lat/long.

Cell regions are based on geographical regions, and some valid cell
locations (such as islands in Europe) are geocoded as None. The
CellAreaUpdater picks a region for the cell area, which may span
multiple regions.  The most common region is picked, and in the case of
a tie, the lexigraphically least could be picked.

If there are an equal number of cells with region=None and a known
region, a TypeError is raised when attempted to sort by strings and
None. In Python 2.x, sorting was allowed, with None sorted as less than
all other strings.