Improve webapp startup debugging
This improves the ability to debug the webapp startup in server
environments where startup issues are masked by gunicorn running the
webapp in a separate worker process. Now when the webapp container is
started, it runs a "check" to see if things can start up, and then
proceeds to run the webapp in gunicorn.