Use pytest.mark.parametrize instead of for loops
Use pytest.mark.parametrize instead of a for loop for the two main types
of observation reports.

This dedents the test code, which can mostly be hidden if whitespace is
ignored, but some lines are also reformatted with Black.

There are no test changes except for test_new, which now expects one set
of metrics per test run instead of two sets, one for each report type.