Add tests for retrying after database exceptions
Add new tests for the two retriable database exceptions.

There is an existing test that sets a short lockout timeout to test the
exception handler for LOCK_WAIT_TIMEOUT. This strategy can't be extended
to other exceptions.

This work exposed an issue with sessions and tests, which the existing
test (unwittingly?) relies on. The connection used by tasks is the same
one used by the test code to setup the database, and a rollback sent
inside the task can cause the test data rows to be rolled back as well.
Raising the retryable exception causes the just-setup cell row to be
dropped, so a new cell is created from the observation. This is probably
not what was intended, but is tested in several ways.

If the task session is now expected to commit data, it is wrapped in a
nested transaction, so that rollbacks don't effect tests. This keeps
the "test" cell for the new retriable tests, so it is an update, not a
new cell.

This nested transaction should work at runtime as well.