Restrict begin_nested to tests
In the task session, run .begin_nested() if explicitly requested by the
test code, and avoid it otherwise.

The test test_retriable_exceptions will not fully clean up after itself
unless the task session runs with a savepoint. However, the rest of the
tests and the application run correctly without it.