Use a shared session for tests
When running tests, use scoped_session to ensure that test code, app
code, and task code all work with the same session. This allows
session.begin_nested() to be used when testing code that calls
session.rollback(). This is currently one test, the recently refactored
test for deadlocks in StationUpdater, but more are expected as we add
more deadlock handlers.

There are a few changes to tests when using a shared, thread-local
session. The first is that session.close() should only be called once at
the end of tests, rather than once for each provisioned session. The
function-scoped session fixture handles closing the session, and a new
function db.release_session(session) can be used instead of
session.close().

Because the session fixture is in charge of session teardown, it must be
included if a new session is provisioned by tested code, or else the
session will remain allocated and cause an error in the next test that
uses a session. If db.session() is called without the session fixture,
an exception is raised, so that the correct test fails.