Expand retry decorator to more updaters
Move _retry_updates_on_mysql_lock_fail to retry_on_mysql_lock_fail in
ichnaea.db, move station-specific stuff to parameters, and implement as
a standard decorator.

This makes it a little more awkward to use in StationUpdater, but allows
it to be expanded to cell area and datamap updating.

The data.station.dberror metric may have tag type=cellarea for retryable
exceptions in CellAreaUpdater, and a new metric datamaps.dberror is
incremented from DataMapUpdater.