Observation docs: center on reports, not signal
The Ichnaea station model attempts to determine the center point and
the area range from which a station will be observed. This updates
the document to reflect this, rather than suggest it is attempting
to locate the station signal source. It includes a diagram of the
observation update process, and the relationship between the model
and the real-world station for some common use cases.

There are additional minor updates from a full read-through of the
document.