bug 1613233: Drop reports with null positions
In the InternalExporter, ensure that a position of null/None is dropped
without an error. For other exporters, confirm that these reports are
forwarded without errors.

In v2/geosubmit, the position item is a dictionary that includes the
latitude and longitude, position source, and other details. It is not
required by the geosubmit endpoints, and it is possible for clients to
submit WiFi and other data without including a GPS-derived position.