bug 1632621: Ensure valid metrics tags
Use markus.util.generate_tag to create valid metrics tags. This
identified an issue with the homepage tag, which was sent as "path:"
before and "path" with generate tag, and appearing as "path: true" in
InfluxDB. It will now appear as "path:.homepage", and the only path
tag starting with a dot.