bug 1632621: Add api_key_repeat_ip
The return of PFADD is 1 if the cardinality was increased (the value is
new) or 0 if the cardinality is the same (the value is repeated).
Record as "api_key_repeat_ip" which is True of PFADD returns
0, meaning we've (probably) already seen this IP address today, as best
estimated by Redis's HyperLogLog. This replaces "api_key_ip_count",
which was 1 for the first call but 0 for the next calls.

It also reformats the code so that less code is executed in the Redis
pipeline block.