bug 1624687: Take signature of validated data
When calculating api_request_sig, use the validated data, and sort by
dictionary keys, to remove some variance from the requests and model a
caching client.

This does not sort the radio lists (bluetooth, cell, or wifi), so an
equivalent request with the list in a different order will have a
different signature.

Serialization failed for the /v2/geosubmit API when position was
omitted, requiring a small change to that schema.