bug 1635789: Remove PFADD log metrics
The metrics api_key_repeat_ip and api_repeat_response are based on the
return from PFADD, used to estimate the number of daily unique IP
addresses and responses. After further research, this does not signify
if a value is a repeat, but if the estimated cardinality has changed.
This is less and less frequent as the cardinality increases, exceeding
99% at 100 million items. This is not a useful metric to track.

The cardinality of the IP addresses is useful, so the related PFADD call
is retained, without capturing the return. The PFCOUNT return, accurate
to about 1%, is still sent as an operational metric.

The cardinality of unique responses is not useful, so the related PFADD
call is removed. The signature may still be useful in BigQuery analysis,
so we keep it for now as a log metric.