bug 1632621: Add application to the metrics table
Mirror the new organization of the metrics document in the alphabetized
table of metrics. It makes it clearer that some metrics are emitted
somewhere other than what is implied by the name, such as
data.batch.upload emitted by the web app, while other data.* metrics are
emitted by the task app.