Add rate controller to queue monitoring task
The station data backlog is computed while measuring the queue sizes,
and the rate controller parameters are read from Redis. If the rate
controller is enabled, the controller is initialized, and the
previous internal state (if any) is loaded. A new global locate sample
rate is determined and written to Redis, along with the rate controller
state.

Since the rate controller parameters will be set manually, they are
validated, and the rate controller turned off if they are invalid.
New validation is used in the API to read the global sample rate as
well, to make it a little safer, defaulting to 100%.