Update tests for MCC 313, PR
MCC 313 used to be only US (such as 313 000 for Tennessee Wireless),
but it is now used in Puerto Rico as well (313 510 for Puerto Rico
Telecom Company). Use MCC 312 for the test instead, and 313 for an
additional multiple countries test.

Drop the Puerto Rico special test case, since this case is now covered,
and the issue link fails because the issue tracker has been disabled in
the original repository.