Replace ExpiringLRUCache with TTLCache
cachetools.TTLCache is a least-recently-used cache that also has a
time-to-live (TTL) for entries. It seems to have the same function as
repoze.lru.ExpiringLRUCache, however lru has not seen updates in 3 years
and is no longer used by Pyramid.

cachetools has a flexible @cached decorator that can be used to memoize
get_key without the manual cache manipulation. It is not thread-safe by
default, so the code now uses gevent's RLock to synchronize access.