Rewrite asset building as makefile
Rewrite node.make as a proper Makefile that uses modification dates to
determine what actions, if any, to take. Previously, all copy and
compile actions were taken each time the targets were run.

Because the makefile is smarter, `make buildjs` and `make buildcss` are
no longer needed to restrict actions to those file types. Instead, the
top-level command is `make assets` to build all the CSS, JS, and other
assets as needed.

Some changes, such as upgrading uglifyjs, are not detected by make,
so a top-level `make clean-assets` is added.

The documentation is updated to reflect these new commands, and
.gitattributes is updated to hint that GitHub should ignore built and
third-party assets.