Install mapbox-gl and plugin, update npm build
Install mapbox-gl 1.3.0.

There is a 2.0.0 that was released this month, but there are a few
issues. It uses modern JS syntax that uglifyjs is unable to parse.
Initializing a map is now a billable event (which may still be free for
MLS's limited usage). The geocode plugin doesn't work with it yet. We'll
update later, maybe.

npm dedupe fails, because copyfile attempts to run on a directory. This
feels like an npm bug, may be fixed after upgrades.

npm shrinkwrap, when starting with a near-empty file, does not seem to
track dev dependencies. Again, this feels like a bug, but it probably OK
for our usage.