Switch contribution map to mapbox-gl
Implement the contributions map in mapbox-gl. It continues to use the
tile-based contribution overlay. It still has controls for geolocation
and geocoding to move around the map. The base map may load faster,
because it is vector-based, and it now supports smooth scrolling and
tilting.

The new code requires changing the Content Security Policy (CSP), to
allow the new interactions with the Mapbox server.