Switch to version number in Dockerfile
Using lts-slim as the node tag made it difficult to see what node
version is being applied by Dependabot, since the change was just the
image hash. Changing back to the version number puts the burden back on
the maintainers to stay on the LTS version.

In node news, v14.17.0 is the LTS version, v16 is current, and will
become the active LTS version on October 26, 2021.