Refactor get_key and Key
Instead of using SQLAlchemy Core and different paths for 1.3 and 1.4,
refactor get_key to use SQLAlchemy ORM, and Key to load from an ORM
object.

Add Key.as_dict() to export as a plain dict, and use this in
the apikey.py script. When asked to show an unknown API key, print a
message rather than raise an exception.