Circle-CI - Split build/push to machine executor
Split the workflow into two jobs:

- test: Build, lint, and test in a Docker runner
- build-push: Build and push in a machine executor

Commands common to both jobs are factored into commands:

- capture_host_info: Capture host and Docker information in an artifact
- create_version_json: Create Dockerflow's version.json
- login_to_dockerhub: Login if credentials are provided

The docker version is updated from 20.10.6 to 20.10.11, the latest.

A filter is used to restrict the build-push job to the main branch and
tags. There's an additional check to keep PR forks from pushing images.

This prepares for building cross-compiled images with buildx, which will
require a machine executor.